폴더 구조화, update README.md
