+++ index.js
@@ -2,10 +2,10 @@ import express from "express";
 import dotenv from "dotenv";
 import articleRoutes from "./routes/articles.js";
 import commentRoutes from "./routes/articleComments.js";
+import products from "./routes/products.js";
 import productCommentRoutes from "./routes/productComments.js";
 import mongoose from "mongoose";
 import cors from "cors";
-import products from "./routes/products.js";
 
 dotenv.config();
 
@@ -17,7 +17,7 @@ app.use(express.json());
 
 app.use("/articles", articleRoutes);
 app.use("/articleComments", commentRoutes);
-app.use("/products", productCommentRoutes);
+app.use("/productComments", productCommentRoutes);
 app.use("/products", products);
 
 const mongoURI = process.env.MONGO_URI;
